More linter satisfaction, speculative fix to seat hijacking
Slowly but surely we're making the linter a little more happy. It'll be a while yet.
Also an untested fix to prevent seat hijacking, there is a specific case that you can boot someone out of a seat, and it is that the vehicle must be yours. Otherwise, its first come, first served.